+++ core/src/test/java/com/ensody/reactivestate/SomeData.kt
@@ -1,3 +1,5 @@
 package com.ensody.reactivestate
 
 internal data class SomeData(val name: String)
+
+internal data class Counter(var count: Int)